Fix classes

+++ src/pl/zsgornik/Ocena.java
@@ -44,7 +44,7 @@ public class Ocena {
         return comment;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
